Move `InstantVectorOperatorBuffer` to its own file.
